Crate call

Running

Running this fuzzer can be done with cargo hfuzz run call. honggfuzz CLI options can be used by setting HFUZZ_RUN_ARGS, such as -n 4 to use 4 threads.

Debugging a panic

Once a panic is found, it can be debugged with cargo hfuzz run-debug per_thing_rational hfuzz_workspace/call/*.fuzz.

Structs

RewardAgent 🔒

Constants

ERA 🔒

MAX_ED_MULTIPLE 🔒

MIN_ED_MULTIPLE 🔒

REWARD_AGENT_ACCOUNT 🔒

Functions

fund_account 🔒

main 🔒

random_call 🔒

random_ed_multiple 🔒

random_existing_pool 🔒

random_signed_origin 🔒

Grab random accounts, either known ones, or new ones.